jpx-dashboard/requirements.txt
# JPX Dashboard — Python Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
tabulate>=0.9.0
